Update ExtendedMath.json
Change the ToFixed() function to return a string to match the javascript function that it emulates (and uses internally).
Previously, this function returned a number which meant that whole numbers would not display the right number of zeroes after the decimal point.

This update could break games that use this function AND manually update this extension.