Bug Fix and Enhancement
Bug Fix: The onScenePostEvents was not resetting the pressed state to false for unpressed buttons which made the C_button_released call return true always after releasing a button once.
Enhancement: Updated the onSceneFirstLoaded and onScenePostEvents to keep track of the last activated controller id. Also added a condition to get the last activated controller id. This makes it possible to dynamically switch between gamepads and just use whichever controller the user picks up instead of having to designate a static controller/player id.
Sorry for the whitespace differences... I'm not a big github user.