Add "Start dragging object" to enable custom picking of object to be dragged
Default dragging requires the cursor to be over an object. This is not great for fabrics of small objects, because there is a lot of space that will not trigger a drag to start.  This change adds a new action to manually start dragging, so game creators can choose when and which objects to grab.

To enable this logic, I added a new option to the Mouse Button property: "None (disable automatic dragging)".
Do you like this implementation, or should I make a separate property, such as: "Enable automatic dragging"?