Add "Start dragging object" to enable custom picking of object to be dragged (#872)
* Add "Start dragging object" to enable custom picking of objects to be dragged

By default, dragging requires the cursor to be over an object. This is not great for fabrics of small objects, because there is a lot of space that will not trigger a drag to start.  This change adds a new action to manually start dragging, so game creators can choose when and which objects to grab.