[Reviewed] [Object Spawner] Added "Random Position" checkbox and deprecated "Offset" properties (#1115)
* Added "Random Position" boolean and deprecated "Offset" properties

Offsets were originally designed to enable random positions, but this new implementation is much cleaner.
The offset method can still be used by adding a value to the position of objects as a subevent to the "Periodically spawn objects".  This was not possible before GDevelop started picking objects created by extensions.